pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "GraphReasoning"
version = "0.3.1"
description = "HypergraphReasoning packages: Use LLM to reason over hypergraphs, combined with multi-agent modeling."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "Markus J. Buehler" },
  { name = "Isabella Stewart", email = "mbuehler@mit.edu" }
]
classifiers = [
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10"
]
dependencies = [
  "numpy",
  "networkx",
  "matplotlib",
  "pandas",
  "transformers",
  "powerlaw",
  "markdown2",
  "pdfkit",
  "bitsandbytes",
  "peft",
  "accelerate",
  "torch",
  "torchvision",
  "torchaudio",
  "huggingface_hub",
  "langchain",
  "langchain-community",
  "openai",
  "pyvis",
  "yachalk",
  "pytesseract",
  "llama-index-embeddings-huggingface",
  "tqdm",
  "ipython",
  "scikit-learn",
  "scipy",
  "seaborn",
  "uuid",
  "pdfminer.six",
  "python-louvain",
  "wkhtmltopdf",
  "autogen",
  "ipykernel",
  "nbformat",
  "nbconvert",
  "langchain-openai>=1.1.12",
  "httpx>=0.28.1",
  "dotenv>=0.9.9",
  "hypernetx>=2.4.0",
  "umap-learn>=0.5.11",
]

[project.urls]
Homepage = "https://github.com/lamm-mit/HyperGraphReasoning"

[tool.setuptools.packages.find]
where = ["."]
include = ["GraphReasoning*"]
exclude = ["build*", "dist*", "notebooks*", "venv*"]
